[김진혁] sprint4

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -3,12 +3,13 @@ package com.sprint.mission.discodeit.service.basic;
 import com.sprint.mission.discodeit.dto.UserCreateRequest;
 import com.sprint.mission.discodeit.dto.UserDto;
 import com.sprint.mission.discodeit.dto.UserUpdateRequest;
-import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
+import com.sprint.mission.discodeit.exception.BusinessLogicException;
+import com.sprint.mission.discodeit.exception.ExceptionCode;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.repository.file.FileUserStatusRepository;
+import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,27 +23,27 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class BasicUserService implements UserService {
     private final UserRepository userRepository;
-    private final FileUserStatusRepository userStatusRepository;
+    private final UserStatusRepository userStatusRepository;
     private final BinaryContentRepository binaryContentRepository;
 
     @Override
     public UserDto createUser(UserCreateRequest request) {
-        String name = request.name();
-        String email = request.email();
-        String password = request.password();
-        if(userRepository.findByName(name).isPresent()) {
-            throw new IllegalArgumentException("이미 존재하는 이름입니다.");
+        String name = request.name().trim();
+        String email = request.email().trim();
+        String password = request.password().trim();
+        if(userRepository.findByUserName(name).isPresent()) {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_EXISTS);
         }
         if(userRepository.findByEmail(email).isPresent()) {
-            throw new IllegalArgumentException("이미 존재하는 이메일입니다. 다른 이메일을 사용해주세요.");
+            throw new BusinessLogicException(ExceptionCode.EMAIL_EXISTS);
         }
-        User newUser = new User(name, email, password);
+        User newUser = new User(name, email, password, request.profileId());
         UserStatus newStatus = new UserStatus(newUser.getId());
         try {
             userRepository.save(newUser);
             userStatusRepository.save(newStatus);
         } catch (Exception e) {
-            throw  new RuntimeException("시스템 오류로 가입에 실패했습니다.");
+            throw  new BusinessLogicException(ExceptionCode.INTERNAL_SERVER_ERROR);
         }
 
         return UserDto.from(newUser, newStatus);
@@ -51,9 +52,9 @@ public class BasicUserService implements UserService {
     @Override
     public UserDto readUser(UUID id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("읽어 오려는 유저가 존재하지 않습니다."));
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
         UserStatus status = userStatusRepository.findByUserId(user.getId())
-                .orElseThrow(() -> new IllegalArgumentException("유저 상태 정보가 없습니다."));
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_STATUS_NOT_FOUND));
 
         return UserDto.from(user, status);
     }
@@ -63,58 +64,61 @@ public class BasicUserService implements UserService {
         List<User> users = userRepository.findAll();
         return users.stream()
                 .map(user -> {UserStatus status = userStatusRepository.findByUserId(user.getId())
-                        .orElseThrow(() -> new IllegalArgumentException("유저가 존재하지 않습니다."));
+                        .orElseGet(() -> new UserStatus(user.getId()));
                 return UserDto.from(user, status);
                 })
                 .collect(Collectors.toList());
     }
 
     @Override
-    public UserDto deleteUser(UUID id, String password) {
+    public void deleteUser(UUID id, String password) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("삭제할 유저가 존재하지 않습니다."));
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
         if(!user.getPassword().equals(password)) {
-            throw new InputMismatchException("패스워드가 일치하지 않습니다.");
+            throw new BusinessLogicException(ExceptionCode.PASSWORD_NOT_MATCH);
         }
-
-        UserStatus status = userStatusRepository.findByUserId(id).orElse(null);
-        UserDto response = UserDto.from(user, status);
 
         if(user.getProfileId() != null) {
             binaryContentRepository.delete(user.getProfileId());
         }
         userStatusRepository.deleteByUserId(id);
         userRepository.delete(id);
-
-        return response;
     }
 
     @Override
     public void updateUser(UUID id, UserUpdateRequest request) {
-        String name = request.name();
-        String email = request.email();
-        String password = request.password();
-        byte[] profileImage = request.profileImage();
-
         User user = userRepository.findById(id)
-                        .orElseThrow(() -> new IllegalArgumentException("검색하신 유저가 존재하지 않습니다."));
-        userRepository.findByName(name)
-                .filter(u -> !u.getId().equals(id))
-                .ifPresent(u -> {throw new IllegalArgumentException("이미 사용중인 이름 입니다.");});
-        userRepository.findByEmail(email)
-                .filter(u -> !u.getId().equals(id))
-                .ifPresent(u -> {throw new IllegalArgumentException("이미 사용중인 이메일 입니다.");});
-        if(profileImage != null) {
-            if(user.getProfileId() != null) {
-                binaryContentRepository.delete(user.getProfileId());
-            }
-            BinaryContent content = new BinaryContent(profileImage);
-
-            binaryContentRepository.save(content);
-            user.setProfileId(content.getId());
+                        .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+        String name = user.getUsername();
+        if(request.name() != null) {
+            userRepository.findByUserName(request.name())
+                    .filter(u -> !u.getId().equals(id))
+                    .ifPresent(u -> {throw new BusinessLogicException(ExceptionCode.MEMBER_EXISTS);});
+            name = request.name();
+        }
+        String email = user.getEmail();
+        if(request.email() != null) {
+            userRepository.findByEmail(email)
+                    .filter(u -> !u.getId().equals(id))
+                    .ifPresent(u -> {throw new BusinessLogicException(ExceptionCode.EMAIL_EXISTS);});
+            email = request.email();
         }
 
-        user.update(name, email, password);
+        String password = (request.password() != null) ? request.password() : user.getPassword();
+
+        UUID profileId = user.getProfileId();
+        if(request.profileId() != null) {
+            if(user.getProfileId() != null && !user.getProfileId().equals(request.profileId())) {
+                try {
+                    binaryContentRepository.delete(user.getProfileId());
+                } catch(IllegalArgumentException e) {
+                    System.out.println("기존 프로필 파일이 이미 존재하지 않아 삭제를 건너뜁니다.");
+                }
+            }
+            profileId = request.profileId();
+        }
+
+        user.update(name, email, password, profileId);
         userRepository.save(user);
     }
 }
